+++ src/bluesearch/utils.py
@@ -26,7 +26,6 @@ import h5py
 import numpy as np
 import pandas as pd
 import spacy
-import tqdm
 
 
 class Timer:
@@ -300,18 +299,19 @@ class H5:
         with h5py.File(h5_path, "r") as f:
             dset = f[dataset_name]
             n_rows = len(dset)
-
             unpop_rows = []
-            iterable = range(0, n_rows, batch_size)
 
-            if verbose:
-                iterable = tqdm.tqdm(iterable)
-
-            for i in iterable:
+            for i in range(0, n_rows, batch_size):
+                if verbose:
+                    print(
+                        f"\rFinding unpopulated rows: {round(100*i/n_rows):>3d}% done",
+                        end="",
+                    )
                 row = dset[i : i + batch_size]
                 is_unpop = np.isnan(row).any(axis=1)  # (batch_size,)
-
                 unpop_rows.extend(list(np.where(is_unpop)[0] + i))
+
+            print("\rFinding unpopulated rows: 100% done", end="")
 
         return np.array(unpop_rows)
 
@@ -404,17 +404,15 @@ class H5:
             final_res_l = []
 
             n_indices = len(sorted_indices)
-            iterable = range(0, n_indices, batch_size)
 
-            if verbose:
-                iterable = tqdm.tqdm(iterable)
-
-            for i in iterable:
+            for i in range(0, n_indices, batch_size):
+                if verbose:
+                    print(f"\rLoading H5: {round(100*i/n_indices):>3d}% done", end="")
                 subarray = dset[sorted_indices[i : i + batch_size]]  # (batch_size, dim)
                 final_res_l.append(subarray)
 
             final_res = np.concatenate(final_res_l, axis=0)
-
+            print("\rLoading H5: 100% done", end="")
             return final_res[unargsort]
 
     @staticmethod